TRY to merge MainWindow + ShellPage
