+++ AutoDarkModeApp/MainWindow.xaml.cs
@@ -1,32 +1,72 @@
 ﻿using System.Diagnostics;
 using AutoDarkModeApp.Contracts.Services;
-using AutoDarkModeApp.Helpers;
+using AutoDarkModeApp.ViewModels;
+using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml;
-using Windows.UI.ViewManagement;
+using Microsoft.UI.Xaml.Input;
+using Windows.System;
 
 namespace AutoDarkModeApp;
 
 public sealed partial class MainWindow : WindowEx
 {
-    private readonly Microsoft.UI.Dispatching.DispatcherQueue dispatcherQueue;
+    public ShellViewModel ViewModel { get; set; }
 
-    private readonly UISettings settings;
-
-    public MainWindow()
+    public MainWindow(ShellViewModel viewModel)
     {
+        ViewModel = viewModel;
         InitializeComponent();
 
         AppWindow.SetIcon(Path.Combine(AppContext.BaseDirectory, "Assets/AutoDarkModeIcon.ico"));
         AppWindow.SetTaskbarIcon(Path.Combine(AppContext.BaseDirectory, "Assets/AutoDarkModeIcon.ico"));
         AppWindow.SetTitleBarIcon(Path.Combine(AppContext.BaseDirectory, "Assets/AutoDarkModeIcon.ico"));
-        Content = null;
 
+        // TODO: Set the title bar icon by updating /Assets/WindowIcon.ico.
+        // A custom title bar is required for full window theme and Mica support.
+        // https://docs.microsoft.com/windows/apps/develop/title-bar?tabs=winui3#full-customization
+        ExtendsContentIntoTitleBar = true;
+        SetTitleBar(TitleBar);
         Title = Debugger.IsAttached ? "Auto Dark Mode Dev" : "Auto Dark Mode";
 
-        Closed += MainWindow_Closed;
+#if DEBUG
+        TitleBar.Subtitle = "Dev";
+#endif
 
-        // Theme change code picked from https://github.com/microsoft/WinUI-Gallery/pull/1239
-        dispatcherQueue = Microsoft.UI.Dispatching.DispatcherQueue.GetForCurrentThread();
+        Closed += MainWindow_Closed;
+    }
+
+    private void NavViewTitleBar_BackRequested(Microsoft.UI.Xaml.Controls.TitleBar sender, object args)
+    {
+        if (NavigationFrame.CanGoBack)
+        {
+            NavigationFrame.GoBack();
+        }
+    }
+
+    private void NavViewTitleBar_PaneToggleRequested(Microsoft.UI.Xaml.Controls.TitleBar sender, object args)
+    {
+        NavigationViewControl.IsPaneOpen = !NavigationViewControl.IsPaneOpen;
+    }
+
+    private KeyboardAccelerator BuildKeyboardAccelerator(VirtualKey key, VirtualKeyModifiers? modifiers = null)
+    {
+        var keyboardAccelerator = new KeyboardAccelerator() { Key = key };
+
+        if (modifiers.HasValue)
+        {
+            keyboardAccelerator.Modifiers = modifiers.Value;
+        }
+
+        keyboardAccelerator.Invoked += OnKeyboardAcceleratorInvoked;
+
+        return keyboardAccelerator;
+    }
+
+    private void OnKeyboardAcceleratorInvoked(KeyboardAccelerator sender, KeyboardAcceleratorInvokedEventArgs args)
+    {
+        var navigationService = App.GetService<INavigationService>();
+        var result = navigationService.GoBack();
+        args.Handled = result;
     }
 
     private async void MainWindow_Closed(object sender, WindowEventArgs args)
@@ -43,7 +83,9 @@ public sealed partial class MainWindow : WindowEx
         });
         //Debug.WriteLine("Save size: " + postion.X + "\t" + postion.Y + "\t" + bounds.Width + "\t" + bounds.Height);
 
-        Close();
+        // TODO: Maybe we could delete it `Close()`. The reason why `Process.GetCurrentProcess().Kill();` is used, it because MapLocationFinder;
+        //Close();
+
         //Debug.WriteLine("Will kill process");
         try
         {
