+++ AutoDarkModeApp/ViewModels/ShellViewModel.cs
@@ -10,9 +10,6 @@ namespace AutoDarkModeApp.ViewModels;
 public partial class ShellViewModel : ObservableRecipient
 {
     [ObservableProperty]
-    private bool isBackEnabled;
-
-    [ObservableProperty]
     private object? selected;
 
     public INavigationService NavigationService
@@ -34,8 +31,6 @@ public partial class ShellViewModel : ObservableRecipient
 
     private void OnNavigated(object sender, NavigationEventArgs e)
     {
-        IsBackEnabled = NavigationService.CanGoBack;
-
         if (e.SourcePageType == typeof(SettingsPage))
         {
             Selected = NavigationViewService.SettingsItem;
